Updated Users Controller Edit
Might need to go back on this. Or just need to make sure that evrything is copied over from DB to user. Not sure how clunky that is, instead of just editing the parts of user that I want to unlike if I want to change the name, I am hoping that it does not change the liked posts if they dont also come along for the round trip